requirements.txt
numpy<2.0.0
torch
pyyaml
tqdm
pydantic>=2.0.0
typing_extensions>=4.0.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.2.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
tavily-python>=0.3.0
termcolor
python-dotenv
pandas

# Development tools
black
isort
pytest